requirements.txt
attrs
brotlipy==0.7.0
certifi==2020.12.5
selenium==3.141.0